Make sklearn.impute a subpackage
